Development

+++ src/main/java/com/codecool/onlineshop/service/AdminService.java
@@ -2,11 +2,11 @@ package com.codecool.onlineshop.service;
 
 import java.util.List;
 
-import com.codecool.onlineshop.dao.OrdersDaoImpl;
-import com.codecool.onlineshop.dao.OrdersHistoryDaoImpl;
-import com.codecool.onlineshop.dao.ProductDao;
-import com.codecool.onlineshop.dao.ProductsDaoImpl;
+import com.codecool.onlineshop.dao.*;
 import com.codecool.onlineshop.model.Order;
+import com.codecool.onlineshop.model.OrderStatus;
+import com.codecool.onlineshop.model.User;
+import com.codecool.onlineshop.model.UserType;
 import com.codecool.onlineshop.view.View;
 
 public class AdminService {
@@ -15,12 +15,18 @@ public class AdminService {
     private ProductsDaoImpl productDao;
     private OrdersDaoImpl orderDao;
     private OrdersHistoryDaoImpl orderHistoryDao;
+    private UsersDaoImpl usersDao;
+    private boolean wrongInput;
+    private final String COLUMNAMOUNT = "Amount";
+    private final String COLUMNPRICE = "Price";
+    
 
     public AdminService() {
         view = new View();
         productDao = new ProductsDaoImpl();
         orderDao = new OrdersDaoImpl();
         orderHistoryDao = new OrdersHistoryDaoImpl();
+        usersDao = new UsersDaoImpl();
     }
     
     public void createNewProduct() {
@@ -29,47 +35,78 @@ public class AdminService {
         view.showMessage(view.ENTERCATEGORY);
         String category = view.getStringInput();
         view.showMessage(view.ENTERPRICE);
-        String price = view.getStringInput();
+        int price = view.getIntegerInput();
         view.showMessage(view.ENTERAMOUNT);
-        String amount = view.getStringInput();
+        int amount = view.getIntegerInput();
         productDao.addNewProduct(name, category, price, amount);
+        productDao = new ProductsDaoImpl();
     }
 
     public void deleteProductAdmin() {
-        view.showMessage(view.DELETEPRODUCT);
-        String productID = view.getStringInput();
-        productDao.deleteProductAdmin(productID);;
+        boolean editProduct = true;
+        while (editProduct) {
+            int productID = getID(view.DELETEPRODUCT);
+            if (productDao.validID(productID)) {
+                productDao.deleteProductAdmin(productID);
+                editProduct = false;
+            } else {
+                view.showMessage(view.WRONGID);
+            }
+        }
     }
 
-    public String getID(String message) {
+    public Integer getID(String message) {
         view.showMessage(message);
-        String productID = view.getStringInput();
+        int productID = view.getIntegerInput();
         return productID;
     }
 
-    public String editValueOfProduct(String message) {
+    public Integer editValueOfProduct(String message) {
+        view.showMessage(message);
+        int producValue = view.getIntegerInput();
+        return producValue;
+    }
+
+    public String editNameOfProduct(String message) {
         view.showMessage(message);
         String producValue = view.getStringInput();
         return producValue;
     }
 
     public void editProductName() {
-        String getProductId = getID(view.ENTERPRODUCTID);
-        String productName = editValueOfProduct(view.ENTERNAME);
-        productDao.editProductName(getProductId, productName);
+        boolean editProduct = true;
+        while(editProduct) {
+            int productId = getID(view.ENTERPRODUCTID);
+            if (productDao.validID(productId)) {
+                String productName = editNameOfProduct(view.ENTERNAME);
+                productDao.editProductName(productId, productName);
+                editProduct = false;
+            } else {
+                view.showMessage(view.WRONGID);
+            }
+        }
     }
 
+    private void editProductValue(String column, String sentence) {
+        boolean editProduct = true;
+        while(editProduct) {
+            int productId = getID(view.ENTERPRODUCTID);
+            if (productDao.validID(productId)) {
+                int productValue = editValueOfProduct(sentence);
+                productDao.editProduct(productId, productValue, column);
+                editProduct = false;
+            } else {
+                view.showMessage(view.WRONGID);
+            }
+        }
+    }
 
     public void editProductPrice() {
-        String getProductId = getID(view.ENTERPRODUCTID);
-        String productPrice = editValueOfProduct(view.ENTERPRICE);
-        productDao.editProductPrice(getProductId, productPrice);
+        editProductValue(COLUMNPRICE, view.ENTERPRICE);
     }
 
     public void  editProductAmount() {
-        String getProductId = getID(view.ENTERPRODUCTID);
-        String productAmount = editValueOfProduct(view.ENTERAMOUNT);
-        productDao.editProductAmount(getProductId, productAmount);
+        editProductValue(COLUMNAMOUNT, view.ENTERAMOUNT);
     }
 
     public void displayAllProductsInShop() {
@@ -89,5 +126,135 @@ public class AdminService {
         List<Order> ordersHistory = orderHistoryDao.getOrderHistoryDetails();
         view.showMessage("Orders History:");
         view.ordersHistoryTable(ordersHistory);
+    }
+
+    public void displayAllCustomers(){
+        usersDao = new UsersDaoImpl();
+        List<User> allUsers = usersDao.getUserData();
+        view.showMessage("Table of all customers");
+        view.usersTable(allUsers);
+    }
+
+    public void deleteUser() {
+        usersDao = new UsersDaoImpl();
+        view.showMessage(view.ENTERUSERID);
+        int userID = view.getIntegerInput();
+        if (usersDao.isValid(userID)) {
+            usersDao.deleteUser(userID);
+            view.showMessage("Successfully Deleted");
+        } else {
+            view.showMessage(view.NOOPTION);
+        }
+    }
+
+    public void updateUserDetails() {
+        usersDao = new UsersDaoImpl();
+        wrongInput = true;
+        view.showMessage(view.ENTERUSERID);
+        int userID = view.getIntegerInput();
+        if (usersDao.isValid(userID)) {
+            view.displayUpdateUserDetailsMenu();
+            while (wrongInput) {
+                handleUpdateUserDetails(userID);
+            }
+            view.showMessage(view.UPDATED);
+        } else {
+            view.showMessage(view.NOOPTION);
+        }
+    }
+
+    public void updateOrderStatus() {
+        orderDao = new OrdersDaoImpl();
+        wrongInput = true;
+        view.showMessage("Enter ID of an Order: ");
+        int orderId = view.getIntegerInput();
+        if (orderDao.isValid(orderId)) {
+            view.displayUpdateStatusMenu();
+            while (wrongInput) {
+                handleOrderStatusUpdate(orderId);
+            }
+            view.showMessage(view.UPDATED);
+        } else {
+            view.showMessage(view.NOOPTION);
+        }
+    }
+
+    private void handleOrderStatusUpdate(int orderId) {
+        int chosenOption = view.getIntegerInput();
+        switch (chosenOption) {
+            case 1:
+                orderDao.updateOrderStatus(orderId, OrderStatus.PENDING.name());
+                wrongInput = false;
+                break;
+            case 2:
+                orderDao.updateOrderStatus(orderId, OrderStatus.SENT.name());
+                wrongInput = false;
+                break;
+            case 3:
+                orderDao.updateOrderStatus(orderId, OrderStatus.DELIVERED.name());
+                wrongInput = false;
+                break;
+            case 4:
+                orderDao.updateOrderStatus(orderId, OrderStatus.CANCELED.name());
+                wrongInput = false;
+                break;
+            case 0:
+                wrongInput = false;
+                break;
+            default:
+                view.showMessage(view.NOOPTION);
+                break;
+        }
+    }
+
+    private void handleUpdateUserDetails(int userID) {
+        String newValue;
+        int chosenOption = view.getIntegerInput();
+        switch (chosenOption) {
+            case 1:
+                newValue = getNewValueFromUser();
+                usersDao.updateUser(userID, "Name", newValue);
+                wrongInput = false;
+                break;
+            case 2:
+                newValue = getNewValueFromUser();
+                usersDao.updateUser(userID, "Password", newValue);
+                wrongInput = false;
+                break;
+            case 3:
+                newValue = handleUserTypeUpdate();
+                usersDao.updateUser(userID, "UserType", newValue);
+                wrongInput = false;
+                break;
+            case 0:
+                wrongInput = false;
+                break;
+            default:
+                view.showMessage(view.NOOPTION);
+                break;
+        }
+    }
+
+    private String handleUserTypeUpdate() {
+        wrongInput = true;
+        while (wrongInput) {
+            view.showMessage("\n1. Customer\n2. Admin");
+            int chosenOption = view.getIntegerInput();
+            switch (chosenOption) {
+                case 1:
+                    return UserType.CUSTOMER.name();
+                case 2:
+                    return UserType.ADMIN.name();
+                default:
+                    view.showMessage(view.NOOPTION);
+                    break;
+            }
+        }
+        return null;
+    }
+
+    private String getNewValueFromUser() {
+        view.showMessage(view.ENTERNEWVALUE);
+        return view.getStringInput();
     }
 }
